refactor: core 폴더 다시 삭제
