+++ src/main/kotlin/com/wheretopop/domain/building/Building.kt
@@ -1,4 +1,4 @@
-package com.wheretopop.domain.building.core
+package com.wheretopop.domain.building
 
 import com.wheretopop.shared.model.Location
 import java.time.Instant
