+++ src/main/kotlin/com/wheretopop/domain/building/BuildingCriteria.kt
@@ -1,4 +1,4 @@
-package com.wheretopop.domain.building.core
+package com.wheretopop.domain.building
 
 
 /**
